refactor: populate OperationSelector programatically
The OperationSelector component used to be populated manually, with four
OperatorCheck components. It also handled the setting of the operators
state on Game component. It now uses an array declared inside of itself,
with the labels for the checkboxes, and uses the map method to populate
with checkboxes and labels, through the OperationCheckBox component,
which is also named more verbosally for clarity. The setting of the game
will be left to the App component on the press of the 'start button'