refactor: convert input setters into components
The Settings component used to have an OperationSelector component and
two different divs with an input and button each. These divs are now
substituted for two SetterInput components, one for the time, and one
for the desired maximum value. This makes the Settings component code
much clearer and easier to maintain. The setting of the state will be
left to the App component on the pressing of the 'Start button'